Move pick and place config parsing out of dev/
No functional changes.
