+++ examples/kuka_iiwa_arm/dev/pick_and_place/README.md
@@ -23,7 +23,7 @@ the processes.
 ┌──────────▶│         ├─────────┐      ┌───────────┐
 │           │         │         │      │           │
 │ ┌────────▶│ Planner │         │      │           │
-│ │         │         │plan     │      │           │            
+│ │         │         │plan     │      │           │
 │ │ ┌──────▶│         ├─────┐   │      │           │ vis info   ┌────────────┐
 │ │ │       └─────────┘     │   └─────▶│           ├───────────▶│ Visualizer │
 │ │ │                       │          │           │            └────────────┘
@@ -71,7 +71,7 @@ various configurations:
 - Select "Scripts" > "2.stop-simulation" from the menu bar to stop any running
   simulations, but leave the visualizer open.  You can also start and stop
   individual processes or groups of processes through the right-click menu of
-  the corresponding line in the GUI. 
+  the corresponding line in the GUI.
 
 # Modifying scenarios
 Double-clicking on one of the process lines in the GUI will bring up a dialog
@@ -80,7 +80,7 @@ process. If you look at the "0.simulators" and "1.planners" groups, you will
 find that the `lcm_pick_and_place_simulator` and `lcm_pick_and_place_planner`
 commands both take a `--configuration_file` flag. For example, the
 configuration file for the single-iiwa scenario is
-`drake/examples/kuka_iiwa_arm/dev/pick_and_place/configuration/single_iiwa.pick_and_place_configuration`.
+`drake/examples/kuka_iiwa_arm/pick_and_place/configuration/single_iiwa.pick_and_place_configuration`.
 Because this file-path begins with `drake/`, the executables will look for it
 under the Drake root directory. You can also supply your own configuration file
 from outside the Drake file hierarchy; in that case, you will need to provide
@@ -88,5 +88,4 @@ the absolute path to the file. To create your custom scenario, simply copy one
 of the existing configuration files and make modifications as desired. The
 configuration files are text-format protobuf files that use the message types
 defined in `<drake
-root>/examples/kuka_iiwa_arm/dev/pick_and_place/pick_and_place_configuration.proto`. 
-
+root>/examples/kuka_iiwa_arm/pick_and_place/pick_and_place_configuration.proto`.
