+++ examples/kuka_iiwa_arm/pick_and_place/configuration/single_iiwa.pick_and_place_configuration
@@ -1,29 +1,29 @@
 # This configuration file defines a scenario in which one Kuka iiwa arm is
-# surrounded by five yellow barrier posts and a low table.  A 6 cm cube is
-# initially located on top of one of the posts. For more details on the format
-# of this file, see
-# <drake root>/examples/kuka_iiwa_arm/dev/pick_and_place/pick_and_place_configuration.proto
+# surrounded by four tables (two high, round tables, and two low rectangular
+# tables). A robot toy is initially located on top of one of the tables. For
+# more details on the format of this file, see
+# <drake root>/examples/kuka_iiwa_arm/pick_and_place/pick_and_place_configuration.proto
 
 # Model Definitions
 # This section defines the models that will be used in the rest of the
 # configuration file.
 model {
   items {
-    key: "cube"
-    value: {
-      simulation_model_path: "drake/examples/kuka_iiwa_arm/models/objects/simple_cuboid.urdf"
-    }
-  }
-  items {
-    key: "yellow_post"
+    key: "big_robot_toy"
     value {
-      simulation_model_path: "drake/examples/kuka_iiwa_arm/models/objects/yellow_post.urdf"
+      simulation_model_path: "drake/examples/kuka_iiwa_arm/models/objects/big_robot_toy.urdf"
     }
   }
   items {
-    key: "extra_heavy_duty_table"
+    key: "round_table"
     value: {
-      simulation_model_path: "drake/examples/kuka_iiwa_arm/models/table/extra_heavy_duty_table_surface_only_collision.sdf";
+      simulation_model_path: "drake/examples/kuka_iiwa_arm/models/objects/round_table.urdf"
+    }
+  }
+  items {
+    key: "folding_table"
+    value: {
+      simulation_model_path: "drake/examples/kuka_iiwa_arm/models/objects/folding_table.urdf"
     }
   }
   items {
@@ -39,75 +39,40 @@ model {
 # each entry defines the pose of the model's base link relative to the world
 # frame.
 table {
-  model_name: "yellow_post"
+  model_name: "round_table"
   optitrack_info {
-    id: 2
-    X_MF {
-      xyz: [0.0, 0.0, 1.02]
-    }
+    id: 8
   }
   pose {
-    xyz: [0.0, 1.0, 0.0]
+    xyz: [0.1, 0.9, 0.91]
   }
 }
 table {
-  model_name: "yellow_post"
+  model_name: "round_table"
   optitrack_info {
-    id: 3
-    X_MF {
-      xyz: [0.0, 0.0, 1.02]
-    }
+    id: 9
   }
   pose {
-    xyz: [0.80, 0.36, 0.0]
+    xyz: [0.8, -0.36, 0.91]
   }
 }
 table {
-  model_name: "extra_heavy_duty_table"
+  model_name: "folding_table"
   optitrack_info {
-    id: 4
-    X_MF {
-      xyz: [0.0, 0.0, 0.7645]
-    }
+    id: 12
   }
   pose {
-    xyz: [0.86, -0.36, -0.07]
+    xyz: [0.8, 0.36, 0.66]
+    rpy: [0.0, 0.0, 0.7]
   }
 }
 table {
-  model_name: "yellow_post"
+  model_name: "folding_table"
   optitrack_info {
-    id: 5
-    X_MF {
-      xyz: [0.0, 0.0, 1.02]
-    }
+    id: 13
   }
   pose {
-    xyz: [0.30, -0.9, 0.0]
-  }
-}
-table {
-  model_name: "yellow_post"
-  optitrack_info {
-    id: 6
-    X_MF {
-      xyz: [0.0, 0.0, 1.02]
-    }
-  }
-  pose {
-    xyz: [-0.1, -1.0, 0.0]
-  }
-}
-table {
-  model_name: "yellow_post"
-  optitrack_info {
-    id: 7
-    X_MF {
-      xyz: [0.0, 0.0, 1.02]
-    }
-  }
-  pose {
-    xyz: [-0.47, -0.8, 0.0]
+    xyz: [0.1, -0.9, 0.66]
   }
 }
 
@@ -121,7 +86,7 @@ robot {
     id: 1
   }
   pose {
-    xyz: [0.0, 0.0, 0.7645];
+    xyz: [0.0, 0.0, 0.7645]
   }
 }
 
@@ -130,13 +95,13 @@ robot {
 # field of each entry defines the initial pose of the model's base link
 # relative to the world frame.
 object {
-  model_name: "cube"
+  model_name: "big_robot_toy"
   optitrack_info {
-    id: 8
+    id: 7
   }
   pose {
-    xyz: [0.80, 0.36, 1.05]
-    rpy: [0.0, 0.0, 0.0]
+    xyz: [0.1, 0.9, 1.0]
+    rpy: [0.0, 0.0, 1.571]
   }
 }
