+++ examples/kuka_iiwa_arm/pick_and_place/pick_and_place_configuration.proto
